refactor: pages.js에서 page.js로 파일명 변경
